Add literal retrieval arm tests, fix score decode and GetHealth reuse
Adds tests/query.rs, exercising store/search.rs directly through
WritePath::query: audience scoping, sharing, a real full-text match,
domain/entity-kind/container scoping, and the score-then-id tiebreak.
Address PR #31 review feedback flagging the lane's missing coverage.

Writing the audience-scoping test surfaced a real bug: the combined
ts_rank + similarity score is Postgres `real`, but LiteralResult.score
is f64, so sqlx failed to decode any row with an actual match. Cast to
float8 in the query.

Also has GetHealth's object-store reachability check call
Error::is_currently_unavailable() instead of re-spelling the
Unavailable(_) match, per the same review.

+++ crates/altaird/src/store/search.rs
@@ -110,7 +110,7 @@ pub async fn literal(
     let mut q = CandidateQuery::new(member, LifecycleScope::Active, &entity::columns())
         .select_also(
             "(ts_rank(e.search_vector, plainto_tsquery('english', $?)) \
-             + similarity(e.search_text, $?)) AS altair_score",
+             + similarity(e.search_text, $?))::float8 AS altair_score",
             [Bind::Text(text.to_owned()), Bind::Text(text.to_owned())],
         )
         .select_also(
